created a customerlist component to render the customer.
